+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
 ## [Unreleased]
+### Changed
+- Rename master branch to main. [#37](https://github.com/IN-CORE/pyincore-viz/issues/37)
+
 ### Fixed
 - Auto build of documentation container. [#34](https://github.com/IN-CORE/pyincore-viz/issues/34)